fix(tests): split oversized edge suite; forbidden names; either-registry invalidation; catalog rows
Codex round 6 on #501: (1) the 846-line edge-context suite splits into
health-availability-privacy.test.ts (identity rendering, malformed
sources) and health-availability-selection.test.ts (pools, budgets,
follow-ups) with a shared fixture module, every file under the size
guideline and registered in safe-core; (2) safeDisplayName drops names
carrying forbidden privacy patterns — URLs, hostname-shaped tokens, IPs,
Markdown delimiters, secret-like words — failing closed to the exact ID;
(3) whole-source invalidation applies to the device registry too: one
malformed device row makes coverage Limited; (4) every omitted group
appears in the catalog with its own counts and a label-addressed detail
request before the aggregate line; (5) the device-subcluster truncation
carries the Progressive Detail follow-up like every other path.
