+++ tests/skills/health-availability-edge-context.test.ts
@@ -1,57 +1,11 @@
 import { describe, expect, it } from "vitest";
 
+import { groupFixture } from "./_health-availability-fixtures.js";
 import {
-  type AvailabilityFixture,
-  type ConfigEntry,
   type RegistryRow,
   type StateRow,
   summarizeAvailability,
 } from "./_health-availability-oracle.js";
-
-type GroupInput = {
-  platform: string;
-  count: number;
-  state: string;
-  disabled_by?: string | null;
-  error_reason?: string;
-  error_reason_translation_key?: string;
-};
-
-function groupFixture(groups: GroupInput[]): AvailabilityFixture {
-  const states: StateRow[] = [];
-  const registry: RegistryRow[] = [];
-  const entries: ConfigEntry[] = [];
-  for (const [groupIndex, group] of groups.entries()) {
-    const entryID = `private-entry-${groupIndex}`;
-    entries.push({
-      entry_id: entryID,
-      domain: group.platform,
-      state: group.state,
-      title: `Private ${groupIndex}`,
-      ...(group.disabled_by !== undefined
-        ? { disabled_by: group.disabled_by }
-        : {}),
-      ...(group.error_reason !== undefined
-        ? { error_reason: group.error_reason }
-        : {}),
-      ...(group.error_reason_translation_key !== undefined
-        ? {
-            error_reason_translation_key: group.error_reason_translation_key,
-          }
-        : {}),
-    });
-    for (let rowIndex = 0; rowIndex < group.count; rowIndex += 1) {
-      const entityID = `sensor.private_${groupIndex}_${rowIndex}`;
-      states.push({ entity_id: entityID, state: "unavailable" });
-      registry.push({
-        entity_id: entityID,
-        config_entry_id: entryID,
-        platform: group.platform,
-      });
-    }
-  }
-  return { states, registry, entries, devices: [] };
-}
 
 describe("health availability adversarial edge fixtures", () => {
   it("routes one sorted budgeted group ledger across both owners", () => {
@@ -189,51 +143,6 @@ describe("health availability adversarial edge fixtures", () => {
     expect(report).not.toContain("private-device");
   });
 
-  it("never emits hostile platform or config-entry domain values", () => {
-    const report = summarizeAvailability({
-      states: [{ entity_id: "sensor.private", state: "unavailable" }],
-      registry: [
-        {
-          entity_id: "sensor.private",
-          config_entry_id: "private-entry",
-          platform: "private-host.local",
-        },
-      ],
-      entries: [
-        {
-          entry_id: "private-entry",
-          domain: "token-secret",
-          state: "loaded",
-          title: "Private",
-        },
-      ],
-      devices: [],
-    });
-    expect(report).not.toContain("private-host.local");
-    expect(report).not.toContain("token-secret");
-    // Codex R4: one malformed registry row (unsafe platform) invalidates the
-    // whole registry source — no derived group survives, coverage is limited.
-    expect(report).toContain("Coverage unavailable: entity registry");
-    expect(report).not.toContain("sensor:");
-
-    const unjoinedAttention = summarizeAvailability({
-      states: [],
-      entries: [
-        {
-          entry_id: "private-entry",
-          domain: "token-secret",
-          state: "setup_error",
-          title: "Private",
-        },
-      ],
-      devices: [],
-    });
-    expect(unjoinedAttention).toContain(
-      "unknown integration: setup_error; impact attribution unavailable",
-    );
-    expect(unjoinedAttention).not.toContain("token-secret");
-  });
-
   it("reports omitted device-cluster and entity-state counts", () => {
     const clusterSizes = [5, 4, 3, 2, 1];
     const states: StateRow[] = [];
@@ -253,7 +162,10 @@ describe("health availability adversarial edge fixtures", () => {
       entries: [],
       devices,
     });
-    expect(report).toContain("2 device clusters omitted (3 entity states).");
+    // Codex R6: sibling truncation paths carry Progressive Detail too.
+    expect(report).toContain(
+      "2 device clusters omitted (3 entity states). Request the full device-subcluster list for all clusters; results may have changed (fresh live read).",
+    );
   });
 
   it("uses both sides of restored/current 60 percent thresholds", () => {
@@ -392,430 +304,6 @@ describe("health availability adversarial edge fixtures", () => {
     });
     expect(systemHealth).toContain("Status: Attention.");
     expect(systemHealth).toContain("Next step: Review failed system health.");
-  });
-
-  it("gives integration-attributed tracker groups first-pool selection under budget pressure", () => {
-    const states: StateRow[] = [];
-    const registry: RegistryRow[] = [];
-    const entries: ConfigEntry[] = [];
-    for (let groupIndex = 0; groupIndex < 10; groupIndex += 1) {
-      const entryID = `private-sensor-${groupIndex}`;
-      entries.push({
-        entry_id: entryID,
-        domain: "private_sensors",
-        state: "loaded",
-        title: `Private ${groupIndex}`,
-      });
-      for (let rowIndex = 0; rowIndex < 5; rowIndex += 1) {
-        const entityID = `sensor.private_${groupIndex}_${rowIndex}`;
-        states.push({ entity_id: entityID, state: "unavailable" });
-        registry.push({
-          entity_id: entityID,
-          config_entry_id: entryID,
-          platform: "private_sensors",
-        });
-      }
-    }
-    // Tracker rows attributed to a mobile_app config entry: the group label
-    // is the integration platform, only its MEMBER domains reveal trackers.
-    entries.push({
-      entry_id: "private-tracker",
-      domain: "mobile_app",
-      state: "loaded",
-      title: "Private tracker",
-    });
-    for (const suffix of ["a", "b", "c"]) {
-      const entityID = `device_tracker.private_${suffix}`;
-      states.push({ entity_id: entityID, state: "unavailable" });
-      registry.push({
-        entity_id: entityID,
-        config_entry_id: "private-tracker",
-        platform: "mobile_app",
-      });
-    }
-    const report = summarizeAvailability({
-      states,
-      registry,
-      entries,
-      devices: [],
-    });
-    expect(report).toContain("device_tracker.private_a");
-    expect(report).toContain("device_tracker.private_c");
-    expect(report).toContain("group details omitted");
-  });
-
-  it("prioritizes current rows over restored rows in large-group examples regardless of input order", () => {
-    const states: StateRow[] = [];
-    const registry: RegistryRow[] = [];
-    for (let rowIndex = 0; rowIndex < 7; rowIndex += 1) {
-      states.push({
-        entity_id: `sensor.private_r${rowIndex}`,
-        state: "unavailable",
-        attributes: { restored: true },
-      });
-    }
-    for (let rowIndex = 0; rowIndex < 5; rowIndex += 1) {
-      states.push({
-        entity_id: `sensor.private_c${rowIndex}`,
-        state: "unavailable",
-      });
-    }
-    for (const row of states) {
-      registry.push({
-        entity_id: row.entity_id,
-        config_entry_id: "private-entry",
-        platform: "private_platform",
-      });
-    }
-    const report = summarizeAvailability({
-      states,
-      registry,
-      entries: [
-        {
-          entry_id: "private-entry",
-          domain: "private_platform",
-          state: "loaded",
-          title: "Private",
-        },
-      ],
-      devices: [],
-    });
-    for (const suffix of ["c0", "c1", "c2", "c3", "c4"]) {
-      expect(report).toContain(`  - sensor.private_${suffix}`);
-    }
-    expect(report).not.toContain("sensor.private_r0");
-    expect(report).toContain("total 12, shown 5, omitted 7.");
-  });
-
-  it("selects attention details in failure-priority order under budget pressure", () => {
-    const report = summarizeAvailability(
-      groupFixture([
-        { platform: "nl_a", count: 10, state: "not_loaded" },
-        { platform: "nl_b", count: 10, state: "not_loaded" },
-        { platform: "nl_c", count: 10, state: "not_loaded" },
-        { platform: "nl_d", count: 10, state: "not_loaded" },
-        { platform: "nl_e", count: 10, state: "not_loaded" },
-        { platform: "nl_f", count: 10, state: "not_loaded" },
-        { platform: "urgent", count: 2, state: "setup_error" },
-      ]),
-    );
-    expect(report).toContain("urgent: setup_error; impact 2 entity states");
-    expect(report.match(/; impact \d+ entity states/g)).toHaveLength(5);
-    expect(report.indexOf("urgent: setup_error; impact")).toBeLessThan(
-      report.indexOf("nl_a: not_loaded; impact"),
-    );
-  });
-
-  it("keeps restored-only inventory out of the current pool", () => {
-    const states: StateRow[] = [];
-    const registry: RegistryRow[] = [];
-    const entries: ConfigEntry[] = [];
-    for (let groupIndex = 0; groupIndex < 5; groupIndex += 1) {
-      const entryID = `private-restored-${groupIndex}`;
-      entries.push({
-        entry_id: entryID,
-        domain: `restored_${groupIndex}`,
-        state: "loaded",
-        title: `Private ${groupIndex}`,
-      });
-      for (let rowIndex = 0; rowIndex < 10; rowIndex += 1) {
-        const entityID = `sensor.private_${groupIndex}_${rowIndex}`;
-        states.push({
-          entity_id: entityID,
-          state: "unavailable",
-          attributes: { restored: true },
-        });
-        registry.push({
-          entity_id: entityID,
-          config_entry_id: entryID,
-          platform: `restored_${groupIndex}`,
-        });
-      }
-    }
-    entries.push({
-      entry_id: "private-current",
-      domain: "current_small",
-      state: "loaded",
-      title: "Private current",
-    });
-    for (const suffix of ["a", "b", "c"]) {
-      const entityID = `sensor.private_current_${suffix}`;
-      states.push({ entity_id: entityID, state: "unavailable" });
-      registry.push({
-        entity_id: entityID,
-        config_entry_id: "private-current",
-        platform: "current_small",
-      });
-    }
-    const report = summarizeAvailability({
-      states,
-      registry,
-      entries,
-      devices: [],
-    });
-    expect(report).toContain("current_small: 3 entity states (0 restored)");
-    expect(report).toContain("sensor.private_current_a");
-    expect(report).toContain("5 group details omitted (50 entity states).");
-  });
-
-  it("replaces identities with deterministic per-type aliases in shareable mode", () => {
-    const report = summarizeAvailability({
-      ...groupFixture([{ platform: "one", count: 3, state: "loaded" }]),
-      privacyMode: "shareable",
-    });
-    expect(report).toContain("  - sensor 1");
-    expect(report).toContain("  - sensor 3");
-    expect(report).not.toContain("sensor.private_0_0");
-  });
-
-  it("counts malformed state entity IDs separately and never renders them", () => {
-    const fixture = groupFixture([
-      { platform: "one", count: 2, state: "loaded" },
-    ]);
-    fixture.states.push({
-      entity_id: "sensor.Secret-Host.local",
-      state: "unavailable",
-    });
-    const report = summarizeAvailability(fixture);
-    expect(report).not.toContain("Secret-Host");
-    expect(report).toContain(
-      "Invalid rows: 1 (excluded from reconciliation, never rendered).",
-    );
-    expect(report).toContain("2 entity states: unavailable 2");
-  });
-
-  it("ranks current stateless rows after current findings in examples", () => {
-    const states: StateRow[] = [];
-    const registry: RegistryRow[] = [];
-    for (let rowIndex = 0; rowIndex < 4; rowIndex += 1) {
-      states.push({
-        entity_id: `button.private_b${rowIndex}`,
-        state: "unknown",
-      });
-    }
-    for (let rowIndex = 0; rowIndex < 5; rowIndex += 1) {
-      states.push({
-        entity_id: `sensor.private_s${rowIndex}`,
-        state: "unavailable",
-      });
-    }
-    for (let rowIndex = 0; rowIndex < 3; rowIndex += 1) {
-      states.push({
-        entity_id: `sensor.private_r${rowIndex}`,
-        state: "unavailable",
-        attributes: { restored: true },
-      });
-    }
-    for (const row of states) {
-      registry.push({
-        entity_id: row.entity_id,
-        config_entry_id: "private-entry",
-        platform: "private_platform",
-      });
-    }
-    const report = summarizeAvailability({
-      states,
-      registry,
-      entries: [
-        {
-          entry_id: "private-entry",
-          domain: "private_platform",
-          state: "loaded",
-          title: "Private",
-        },
-      ],
-      devices: [],
-    });
-    for (const suffix of ["s0", "s1", "s2", "s3", "s4"]) {
-      expect(report).toContain(`  - sensor.private_${suffix}`);
-    }
-    expect(report).not.toContain("  - button.private_b0");
-    expect(report).toContain("total 12, shown 5, omitted 7.");
-  });
-
-  it("renders owned entity details for displayed failed integrations", () => {
-    const report = summarizeAvailability(
-      groupFixture([{ platform: "hue", count: 3, state: "setup_error" }]),
-    );
-    expect(report).toContain("hue: setup_error; impact 3 entity states");
-    for (const idx of [0, 1, 2]) {
-      expect(report).toContain(`  - sensor.private_0_${idx}`);
-    }
-  });
-
-  it("orders the tracker pool by finding priority under budget pressure", () => {
-    const states: StateRow[] = [];
-    const registry: RegistryRow[] = [];
-    const entries: ConfigEntry[] = [];
-    for (let groupIndex = 0; groupIndex < 5; groupIndex += 1) {
-      const entryID = `private-tracker-${groupIndex}`;
-      entries.push({
-        entry_id: entryID,
-        domain: `tracker_${groupIndex}`,
-        state: "loaded",
-        title: `Private ${groupIndex}`,
-      });
-      for (let rowIndex = 0; rowIndex < 10; rowIndex += 1) {
-        const entityID = `device_tracker.private_${groupIndex}_${rowIndex}`;
-        states.push({ entity_id: entityID, state: "unavailable" });
-        registry.push({
-          entity_id: entityID,
-          config_entry_id: entryID,
-          platform: `tracker_${groupIndex}`,
-        });
-      }
-    }
-    entries.push({
-      entry_id: "private-urgent",
-      domain: "tracker_urgent",
-      state: "setup_error",
-      title: "Private urgent",
-    });
-    states.push({
-      entity_id: "device_tracker.private_urgent",
-      state: "unavailable",
-    });
-    registry.push({
-      entity_id: "device_tracker.private_urgent",
-      config_entry_id: "private-urgent",
-      platform: "tracker_urgent",
-    });
-    const report = summarizeAvailability({
-      states,
-      registry,
-      entries,
-      devices: [],
-    });
-    expect(report).toContain(
-      "tracker_urgent: setup_error; impact 1 entity states",
-    );
-    expect(report).toContain("  - device_tracker.private_urgent");
-    expect(report).toContain("group details omitted");
-  });
-
-  it("attaches exact follow-ups to every truncation path", () => {
-    const largeGroup = summarizeAvailability(
-      groupFixture([{ platform: "one", count: 12, state: "loaded" }]),
-    );
-    expect(largeGroup).toContain(
-      "total 12, shown 5, omitted 7. Request this group's full detail for all rows; results may have changed (fresh live read).",
-    );
-
-    const manyGroups = summarizeAvailability(
-      groupFixture(
-        Array.from({ length: 11 }, (_, index) => ({
-          platform: `many_${String.fromCharCode(97 + index)}`,
-          count: 10,
-          state: "loaded",
-        })),
-      ),
-    );
-    expect(manyGroups).toMatch(
-      /group details omitted \(\d+ entity states\)\. Request a group's detail by name for its full rows; results may have changed \(fresh live read\)\./,
-    );
-
-    const manyAttention = summarizeAvailability(
-      groupFixture(
-        Array.from({ length: 26 }, (_, index) => ({
-          platform: `attn_${String.fromCharCode(97 + index)}`,
-          count: 1,
-          state: "setup_error",
-        })),
-      ),
-    );
-    expect(manyAttention).toContain(
-      "1 attention entries omitted by integration-entry cap. Request the full integration-entry list for the rest; results may have changed (fresh live read).",
-    );
-  });
-
-  it("renders the safely renderable friendly name beside the exact ID in private mode", () => {
-    const report = summarizeAvailability({
-      states: [
-        {
-          entity_id: "sensor.private_named",
-          state: "unavailable",
-          attributes: { friendly_name: "Private kitchen sensor" },
-        },
-        {
-          entity_id: "sensor.private_hostile",
-          state: "unavailable",
-          attributes: { friendly_name: "line\nbreak" },
-        },
-      ],
-      registry: [
-        {
-          entity_id: "sensor.private_named",
-          config_entry_id: "private-entry",
-          platform: "private_platform",
-        },
-        {
-          entity_id: "sensor.private_hostile",
-          config_entry_id: "private-entry",
-          platform: "private_platform",
-        },
-      ],
-      entries: [
-        {
-          entry_id: "private-entry",
-          domain: "private_platform",
-          state: "loaded",
-          title: "Private",
-        },
-      ],
-      devices: [],
-    });
-    expect(report).toContain("  - sensor.private_named (Private kitchen sensor)");
-    // A control-character name is not safely renderable — ID only.
-    expect(report).toContain("  - sensor.private_hostile\n");
-    expect(report).not.toContain("line\nbreak");
-  });
-
-  it("prioritizes current findings over stateless context in the current pool", () => {
-    const states: StateRow[] = [];
-    const registry: RegistryRow[] = [];
-    const entries: ConfigEntry[] = [];
-    for (let groupIndex = 0; groupIndex < 5; groupIndex += 1) {
-      const entryID = `private-buttons-${groupIndex}`;
-      entries.push({
-        entry_id: entryID,
-        domain: `buttons_${groupIndex}`,
-        state: "loaded",
-        title: `Private ${groupIndex}`,
-      });
-      for (let rowIndex = 0; rowIndex < 10; rowIndex += 1) {
-        const entityID = `button.private_${groupIndex}_${rowIndex}`;
-        states.push({ entity_id: entityID, state: "unknown" });
-        registry.push({
-          entity_id: entityID,
-          config_entry_id: entryID,
-          platform: `buttons_${groupIndex}`,
-        });
-      }
-    }
-    entries.push({
-      entry_id: "private-finding",
-      domain: "finding_small",
-      state: "loaded",
-      title: "Private finding",
-    });
-    for (const suffix of ["a", "b"]) {
-      const entityID = `sensor.private_finding_${suffix}`;
-      states.push({ entity_id: entityID, state: "unavailable" });
-      registry.push({
-        entity_id: entityID,
-        config_entry_id: "private-finding",
-        platform: "finding_small",
-      });
-    }
-    const report = summarizeAvailability({
-      states,
-      registry,
-      entries,
-      devices: [],
-    });
-    expect(report).toContain("finding_small: 2 entity states (0 restored)");
-    expect(report).toContain("  - sensor.private_finding_a");
-    expect(report).toContain("group details omitted");
   });
 
   it("attributes nothing on a config_entry_id without a matching entry", () => {
